Add documentation for problems
