+++ src/problems/mod.rs
@@ -1,62 +1,66 @@
-use std::{marker::PhantomData, ops::Range};
+use std::ops::Range;
 
-use better_any::{Tid, TidAble};
-use rayon::iter::{IntoParallelRefMutIterator, ParallelIterator};
 use trait_set::trait_set;
 
-use crate::{CustomState, State};
-
 pub mod encoding;
+pub mod evaluate;
 pub mod individual;
 pub mod objective;
 
 pub use encoding::AnyEncoding;
+pub use evaluate::{Evaluate, Evaluator, ObjectiveFunction};
 pub use individual::Individual;
 pub use objective::{MultiObjective, Objective, SingleObjective};
 
-/// Metadata of an optimization problem.
+/// An optimization (minimization) problem.
+///
 /// This trait is the base trait for all problems, and itself only defines
-/// - solution encoding ([`Self::Encoding`]),
-/// - objective type ([`Self::Objective`]), and
-/// - the name of the problem ([`Self::name`]).
+/// - a encoding to solutions to the problem ([`Problem::Encoding`]),
+/// - the type of objective to minimize ([`Problem::Objective`]), and
+/// - the name of the problem ([`Problem::name`]).
 ///
-/// # Example
+/// # Examples
 ///
-/// A simple implementation of the one-dimensional real-valued sphere function `f(x) = x^2`:
+/// A simple implementation of the real-valued sphere function `f(x) = x^2`:
 ///
 /// ```
 /// use mahf::{Problem, SingleObjective};
 ///
-/// pub struct Sphere1D;
+/// pub struct Sphere {
+///     pub dim: usize,
+/// }
 ///
-/// impl Problem for Sphere1D {
+/// impl Problem for Sphere {
 ///     type Encoding = Vec<f64>; // real-valued vector
 ///     type Objective = SingleObjective;
 ///
-///     fn name(&self) -> &str { "Sphere1D" }
+///     fn name(&self) -> &str {
+///         "Sphere"
+///     }
 /// }
 /// ```
 pub trait Problem: 'static {
-    /// The encoding of a solution to the optimization problem.
+    /// The encoding of a solution to the optimization problem (genotype).
     type Encoding: AnyEncoding;
 
-    /// The objective type.
-    /// See [`SingleObjective`] and [`MultiObjective`].
+    /// The objective type to minimize.
+    ///
+    /// See [`SingleObjective`] and [`MultiObjective`] for the default options.
     type Objective: Objective;
 
-    /// The name of the problem.
+    /// The name of the optimization problem.
     fn name(&self) -> &str;
 }
 
 trait_set! {
-    /// An optimization problem with a single objective.
+    /// An optimization problem with a single objective or multiple combined objectives.
     ///
     /// This trait should be used in favor over specifying the objective type
     /// of [`Problem`] directly, and is automatically implemented for all problems
     /// with [`SingleObjective`] as [`Problem::Objective`].
     pub trait SingleObjectiveProblem = Problem<Objective = SingleObjective>;
 
-    /// An optimization problem with multiple objectives.
+    /// An optimization problem with multiple distinct objectives.
     ///
     /// This trait should be used in favor over specifying the objective type
     /// of [`Problem`] directly, and is automatically implemented for all problems
@@ -64,124 +68,188 @@ trait_set! {
     pub trait MultiObjectiveProblem = Problem<Objective = MultiObjective>;
 }
 
-pub trait ObjectiveFunction: Problem + Send {
-    fn objective(solution: &Self::Encoding) -> Self::Objective;
-}
-
-pub trait TryDefault: Sized {
-    type Error;
-
-    fn try_default() -> Result<Self, Self::Error>;
-}
-
-impl<T: Default> TryDefault for T {
-    type Error = ();
-
-    fn try_default() -> Result<Self, Self::Error> {
-        Ok(T::default())
-    }
-}
-
-pub trait Evaluate: TryDefault + Send {
-    type Problem: Problem;
-
-    fn evaluate(
-        &mut self,
-        problem: &Self::Problem,
-        state: &mut State<Self::Problem>,
-        individuals: &mut [Individual<Self::Problem>],
-    );
-}
-
-trait_set! {
-    pub trait Evaluator = Evaluate + for<'a> CustomState<'a> + for<'a> TidAble<'a>;
-}
-
-#[derive(Tid)]
-pub struct Sequential<P: ObjectiveFunction + 'static>(PhantomData<P>);
-
-impl<P: ObjectiveFunction> Sequential<P> {
-    pub fn new() -> Self {
-        Self(PhantomData)
-    }
-}
-
-impl<P: ObjectiveFunction> Default for Sequential<P> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<P> Evaluate for Sequential<P>
-where
-    P: ObjectiveFunction,
-{
-    type Problem = P;
-
-    fn evaluate(
-        &mut self,
-        _problem: &Self::Problem,
-        _state: &mut State<Self::Problem>,
-        individuals: &mut [Individual<Self::Problem>],
-    ) {
-        for individual in individuals {
-            individual.evaluate_with(P::objective);
-        }
-    }
-}
-
-impl<P: ObjectiveFunction> CustomState<'_> for Sequential<P> {}
-
-#[derive(Tid)]
-pub struct Parallel<P: ObjectiveFunction + 'static>(PhantomData<P>);
-
-impl<P: ObjectiveFunction> Parallel<P> {
-    pub fn new() -> Self {
-        Self(PhantomData)
-    }
-}
-
-impl<P: ObjectiveFunction> Default for Parallel<P> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<P> Evaluate for Parallel<P>
-where
-    P: ObjectiveFunction + Problem,
-{
-    type Problem = P;
-
-    fn evaluate(
-        &mut self,
-        _problem: &Self::Problem,
-        _state: &mut State<Self::Problem>,
-        individuals: &mut [Individual<Self::Problem>],
-    ) {
-        individuals
-            .par_iter_mut()
-            .for_each(|individual| individual.evaluate_with(P::objective));
-    }
-}
-
-impl<P: ObjectiveFunction> CustomState<'_> for Parallel<P> {}
-
+/// A vector-based optimization problem.
+///
+/// This trait extends the [`Problem`] trait and represents an optimization problem
+/// whose solutions are encoded as vectors with some [`Element`] type.
+///
+/// [`Element`]: VectorProblem::Element
+///
+/// # Examples
+///
+/// A simple implementation of the n-dimensional real-valued sphere function `f(x) = x^2`:
+///
+/// ```
+/// use mahf::{Problem, SingleObjective, problems::VectorProblem};
+///
+/// pub struct Sphere {
+///     pub dim: usize,
+/// }
+///
+/// impl Problem for Sphere {
+///     type Encoding = Vec<f64>;
+///     type Objective = SingleObjective;
+///
+///     fn name(&self) -> &str {
+///         "Sphere"
+///     }
+/// }
+///
+/// impl VectorProblem for Sphere {
+///     type Element = f64;
+///
+///     fn dimension(&self) -> usize {
+///         self.dim
+///     }
+/// }
+/// ```
 pub trait VectorProblem: Problem<Encoding = Vec<Self::Element>> {
+    /// The element type of the vector encoding the solutions.
     type Element: Clone;
 
+    /// Returns the dimension of the optimization problem.
+    ///
+    /// The dimension represents the length of the vector encoding the solutions.
     fn dimension(&self) -> usize;
 }
 
+/// A vector-based optimization problem with limited search space.
+///
+/// This trait extends the [`VectorProblem`] trait and represents an optimization problem
+/// whose solutions are encoded as vectors with a limited search space defined by some domain.
+///
+/// # Examples
+///
+/// A simple implementation of the n-dimensional real-valued sphere function `f(x) = x^2`
+/// restricted to the \[-1, 1) domain on each dimension:
+///
+/// ```
+/// use std::ops::{Range, RangeInclusive};
+/// use mahf::{Problem, SingleObjective, problems::{VectorProblem, LimitedVectorProblem}};
+///
+/// pub struct Sphere {
+///     pub dim: usize,
+/// }
+///
+/// impl Problem for Sphere {
+///     type Encoding = Vec<f64>;
+///     type Objective = SingleObjective;
+///
+///     fn name(&self) -> &str {
+///         "Sphere"
+///     }
+/// }
+///
+/// impl VectorProblem for Sphere {
+///     type Element = f64;
+///
+///     fn dimension(&self) -> usize {
+///         self.dim
+///     }
+/// }
+///
+/// impl LimitedVectorProblem for Sphere {
+///     fn domain(&self) -> Vec<Range<Self::Element>> {
+///         std::iter::repeat(-1.0..1.0).take(self.dim).collect()
+///     }
+/// }
+/// ```
 pub trait LimitedVectorProblem: VectorProblem {
+    /// Returns the bounds of the search space for each element in the vector encoding.
+    ///
+    /// The bounds specify the range of valid values for each element in the vector.
     fn domain(&self) -> Vec<Range<Self::Element>>;
 }
 
+/// A single-objective optimization problem with an indicator for reaching the optimum.
+///
+/// If the objective value of the optimum is known directly, implement [`KnownOptimumProblem`]
+/// instead, which provides a blanket implementation for this trait.
+///
+/// # Examples
+///
+/// A simple implementation of the real-valued sphere function `f(x) = x^2`, where
+/// the optimum is known to be 0.
+///
+/// Note that a implementation of [`KnownOptimumProblem`] is more sensible in this case because
+/// the value is known directly.
+///
+/// ```
+/// use std::ops::Range;
+/// use mahf::{Problem, SingleObjective, problems::OptimumReachedProblem};
+///
+/// pub struct Sphere {
+///     pub dim: usize,
+/// }
+///
+/// impl Problem for Sphere {
+///     type Encoding = Vec<f64>;
+///     type Objective = SingleObjective;
+///
+///     fn name(&self) -> &str {
+///         "Sphere"
+///     }
+/// }
+///
+/// impl OptimumReachedProblem for Sphere {
+///     fn optimum_reached(&self, objective: SingleObjective) -> bool {
+///         // Approximately 0 is accepted as optimum.
+///         objective.value() < 1e-8
+///     }
+/// }
+///
+/// // A value smaller as 1e-8 counts as optimum.
+/// let sphere = Sphere { dim: 1 };
+/// assert!(sphere.optimum_reached(1e-9.try_into().unwrap()));
+/// ```
 pub trait OptimumReachedProblem: SingleObjectiveProblem {
+    /// Checks whether the objective value has reached the optimum.
     fn optimum_reached(&self, objective: SingleObjective) -> bool;
 }
 
+/// A single-objective optimization problem with a known optimum value.
+///
+/// # Examples
+///
+/// A simple implementation of the real-valued sphere function `f(x) = x^2`, where
+/// the optimum is known to be 0.
+///
+/// ```
+/// use std::ops::Range;
+/// use mahf::{Problem, SingleObjective, problems::{KnownOptimumProblem, OptimumReachedProblem}};
+///
+/// pub struct Sphere {
+///     pub dim: usize,
+/// }
+///
+/// impl Problem for Sphere {
+///     type Encoding = Vec<f64>;
+///     type Objective = SingleObjective;
+///
+///     fn name(&self) -> &str {
+///         "Sphere"
+///     }
+/// }
+///
+/// impl KnownOptimumProblem for Sphere {
+///     const DELTA: f64 = 1e-10;
+///
+///     fn known_optimum(&self) -> SingleObjective {
+///         0.0.try_into().unwrap()
+///     }
+/// }
+///
+/// // A value greater than 1e-10 does not counts as optimum.
+/// let sphere = Sphere { dim: 1 };
+/// assert!(!sphere.optimum_reached(1e-9.try_into().unwrap()));
+/// ```
 pub trait KnownOptimumProblem: SingleObjectiveProblem {
+    /// A constant representing the tolerance level for comparing objective values.
+    ///
+    /// This value is used as tolerance to automatically implement [`OptimumReachedProblem`].
+    const DELTA: f64 = 1e-8;
+
+    /// Retrieves the known optimum objective value for the optimization problem.
     fn known_optimum(&self) -> SingleObjective;
 }
 
@@ -189,13 +257,61 @@ impl<P: KnownOptimumProblem> OptimumReachedProblem for P {
     fn optimum_reached(&self, objective: SingleObjective) -> bool {
         let provided = objective.value();
         let known = self.known_optimum().value();
+        debug_assert!(
+            provided >= known,
+            "the provided objective value is smaller than the known optimum"
+        );
 
-        (provided - known) < 1e-8
+        (provided - known).abs() <= P::DELTA
     }
 }
 
+/// The Travelling Salesperson Problem (TSP) optimization problem.
+///
+/// TSP is a single-objective optimization problem that involves finding
+/// the shortest possible route that visits a given set of vertices and
+/// returns to the starting vertex.
+///
+/// # Examples
+///
+/// A simple implementation of the Travelling Salesperson Problem given a edge weight matrix:
+///
+/// ```
+/// use mahf::{Problem, problems::{VectorProblem, TravellingSalespersonProblem}, SingleObjective};
+///
+/// pub struct TSP {
+///     pub instance: String,
+///     pub edge_weights: Vec<Vec<f64>>,
+/// }
+///
+/// impl Problem for TSP {
+///     type Encoding = Vec<usize>;
+///     type Objective = SingleObjective;
+///
+///     fn name(&self) -> &str {
+///         &self.instance
+///     }
+/// }
+///
+/// impl VectorProblem for TSP {
+///     type Element = usize;
+///
+///     fn dimension(&self) -> usize {
+///         self.edge_weights.len()
+///     }
+/// }
+///
+/// impl TravellingSalespersonProblem for TSP {
+///     fn distance(&self, edge: (usize, usize)) -> f64 {
+///         let (source, target) = edge;
+///         self.edge_weights[source][target]
+///     }
+/// }
+/// ```
 pub trait TravellingSalespersonProblem:
     SingleObjectiveProblem + VectorProblem<Element = usize>
 {
+    /// Calculates the distance between two locations,
+    /// i.e. the edge weight between two vertices identified by their indices.
     fn distance(&self, edge: (usize, usize)) -> f64;
 }
